# lib/api-zod/src/generated/types/inquiryStatus.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * ATC catalogue and content workspace API
 * OpenAPI spec version: 0.1.0
 */

export type InquiryStatus = typeof InquiryStatus[keyof typeof InquiryStatus];


export const InquiryStatus = {
  submitted: 'submitted',
  in_progress: 'in_progress',
  responded: 'responded',
} as const;